feat(api): add /me/ai-usage/history + extract AiUsageChart (US-M13.4)
Backend
- New `GET /api/v1/me/ai-usage/history?days=N` (default 30, max 90 via
  `Query(ge=1, le=90)`). Read-only; reuses `AiUsageRepo.get_window` so
  this endpoint and the existing admin metric share window semantics.
- `AiUsageHistoryPoint` Pydantic model in `api/models/user.py`.
- Pytest cases for the zero-state and the days clamp.

Frontend chart extraction
- Lift the hand-rolled `AiUsageChart` SVG from
  `web/src/pages/admin/DashboardPage.tsx` to a shared
  `web/src/components/AiUsageChart.tsx` so the M13.4 user-facing
  Settings → Usage page can reuse it without a chart library.
- Public API: `<AiUsageChart points={...} ariaLabel?={...} />`. Same
  testid (`ai-usage-chart`) and same translation keys
  (`admin:common.loading`, `admin:dashboard.empty`,
  `admin:dashboard.aiUsage.title`) preserved so the existing admin
  Dashboard test stays green without changes.
